+++ src/main/kotlin/domain/DriverDecision.kt
@@ -1,4 +1,4 @@
-package util
+package domain
 
 enum class DriverDecision(val distance: String) {
     GO("-"),
